Move vendor dashboard files into vendor_dashboard directory
