+++ config.php
@@ -1,4 +1,4 @@
 <?php
 session_start();
-require_once __DIR__ . '/db.php';
+require_once __DIR__ . '/vendor_dashboard/db.php';
 ?>
