+++ vendor_dashboard/api/_update_user.php
@@ -1,5 +1,5 @@
 <?php
-require_once __DIR__ . '/../config.php';
+require_once __DIR__ . '/../../config.php';
 header('Content-Type: application/json');
 
 if (!isset($_SESSION['user_id'])) {
