+++ vendor_dashboard/api/delete_file.php
@@ -1,5 +1,5 @@
 <?php
-require_once __DIR__ . '/../config.php';
+require_once __DIR__ . '/../../config.php';
 
 $id = $_POST['id'] ?? null;
 if (!$id) {
