+++ vendor_dashboard/api/upload_file.php
@@ -1,5 +1,5 @@
 <?php
-require_once __DIR__ . '/../config.php';
+require_once __DIR__ . '/../../config.php';
 
 if (!isset($_FILES['pdf'])) {
     http_response_code(400);
